refactor: splitted PatronControllerAsText
Following the report "Search and Reporting", we've refactored
PatronControllerAsText to lower the cohesion. The created classes handle
more specific logic thus applying Single responsibility.
